Integrated with 4.1.6 version of SDK and changed deployment target to 11.0
