--- KountDataCollectorUnivesalLibrary/TouchControls/UIControl+CustomControl.h
@@ -1,33 +1,0 @@
-//
-//  UIControl+CustomControl.h
-//  TouchDemo
-//
-//  Created by Astha Ameta on 19/08/20.
-//  Copyright © 2020 Kount Inc. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface UIButton (CustomButton)
-
-@end
-
-@interface UISlider (CustomSlider)
-
-@end
-
-@interface UIStepper (CustomStepper)
-
-@end
-
-@interface UIPageControl (CustomPageControl)
-
-@end
-
-@interface UIColorWell (CustomColorWell)
-
-@end
-
-@interface UIScrollView (CustomScrollView)
-
-@end
